+++ common/src/main/java/tocraft/walkers/eventhandler/RespawnHandler.java
@@ -3,7 +3,6 @@ package tocraft.walkers.eventhandler;
 import net.minecraft.server.level.ServerPlayer;
 import tocraft.craftedcore.event.common.PlayerEvents;
 import tocraft.walkers.api.PlayerShapeChanger;
-import tocraft.walkers.impl.DimensionsRefresher;
 import tocraft.walkers.impl.PlayerDataProvider;
 
 public final class RespawnHandler implements PlayerEvents.PlayerRespawn {
@@ -14,7 +13,7 @@ public final class RespawnHandler implements PlayerEvents.PlayerRespawn {
         ((PlayerDataProvider) newPlayer).walkers$setAbilityCooldown(((PlayerDataProvider) oldPlayer).walkers$getAbilityCooldown());
         ((PlayerDataProvider) newPlayer).walkers$setRemainingHostilityTime(((PlayerDataProvider) oldPlayer).walkers$getRemainingHostilityTime());
 
-        ((DimensionsRefresher) newPlayer).shape_refreshDimensions();
+        newPlayer.refreshDimensions();
 
         PlayerShapeChanger.sync(newPlayer);
     }