+++ common/src/main/java/tocraft/walkers/impl/DimensionsRefresher.java
@@ -1,5 +1,0 @@
-package tocraft.walkers.impl;
-
-public interface DimensionsRefresher {
-    void shape_refreshDimensions();
-}