+++ common/src/main/java/tocraft/walkers/mixin/PlayerManagerMixin.java
@@ -5,6 +5,7 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.players.PlayerList;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.attributes.AttributeInstance;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -13,7 +14,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import tocraft.walkers.Walkers;
 import tocraft.walkers.api.PlayerShape;
 import tocraft.walkers.api.PlayerShapeChanger;
-import tocraft.walkers.impl.DimensionsRefresher;
 
 @Mixin(PlayerList.class)
 public class PlayerManagerMixin {
@@ -24,27 +24,31 @@ public class PlayerManagerMixin {
         LivingEntity shape = PlayerShape.getCurrentShape(player);
 
         // refresh entity hitbox dimensions after death
-        ((DimensionsRefresher) player).shape_refreshDimensions();
+        player.refreshDimensions();
 
         if (shape != null) {
-            // Re-sync max health for shapes
+            // Re-sync max health for shape
             if (Walkers.CONFIG.scalingHealth) {
                 player.getAttribute(Attributes.MAX_HEALTH)
                         .setBaseValue(Math.min(Walkers.CONFIG.maxHealth, shape.getMaxHealth()));
                 player.setHealth(player.getMaxHealth());
             }
-            // Re-sync amor for shapes
+            // Re-sync amor for shape
             if (Walkers.CONFIG.scalingAmor) {
                 player.getAttribute(Attributes.ARMOR)
                         .setBaseValue(Math.min(Walkers.CONFIG.maxAmor, shape.getAttributeBaseValue(Attributes.ARMOR)));
                 player.getAttribute(Attributes.ARMOR_TOUGHNESS)
                         .setBaseValue(Math.min(Walkers.CONFIG.maxAmorToughness, shape.getAttributeBaseValue(Attributes.ARMOR_TOUGHNESS)));
             }
-            // sync max health & attack damage with clients
-            if (Walkers.CONFIG.scalingHealth || Walkers.CONFIG.percentScalingHealth || Walkers.CONFIG.scalingAmor) {
-                player.connection.send(new ClientboundUpdateAttributesPacket(player.getId(),
-                        player.getAttributes().getSyncableAttributes()));
+            // Re-sync scale for shape
+            AttributeInstance playerScaleAttribute = player.getAttribute(Attributes.SCALE);
+            AttributeInstance shapeScaleAttribute = shape.getAttribute(Attributes.SCALE);
+            if (playerScaleAttribute != null && shapeScaleAttribute != null) {
+                shapeScaleAttribute.setBaseValue(playerScaleAttribute.getBaseValue());
             }
+            // sync max health & attack damage with clients
+            player.connection.send(new ClientboundUpdateAttributesPacket(player.getId(),
+                    player.getAttributes().getSyncableAttributes()));
         }
 
         // send sync packets